Split up some bigger specs and updated the YAML to only do chromium on master
